requirements.txt
streamlit
ultralytics
opencv-python
pillow
streamlit-image-comparison
streamlit-drawable-canvas
pandas